[Refactor] Move survey test functions into base file
This makes functions more resuable and is a first step in attempting to consolidate the testing functions used for surveys and the projects that uses molo.surveys
